Refactor backend Sentry observability
